Extract mapWithConcurrency helper
